+++ handlers/__init__.py
@@ -1,0 +1,2 @@
+from .cli import cli_handler  #noqa #pyright: ignore[reportUnusedImport]
+from .fifo import fifo_handler  #noqa #pyright: ignore[reportUnusedImport]
